dog_posture_detection/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
pandas>=2.0.0
PyYAML>=6.0
scikit-learn>=1.3.0
